+++ lib/core/services/d_reader_wallet_service.dart
@@ -39,7 +39,7 @@ class DReaderWalletService {
 
   Future<String?> listItem({
     required String mintAccount,
-    required double price,
+    required int price,
     String? printReceipt,
   }) async {
     return IoCContainer.resolveContainer<AuctionHouseRepositoryImpl>().listItem(
@@ -53,7 +53,7 @@ class DReaderWalletService {
 
   Future<String?> buyItem({
     required String mint,
-    required double price,
+    required int price,
     required String sellerAddress,
   }) {
     return IoCContainer.resolveContainer<AuctionHouseRepositoryImpl>()
